+++ CHANGELOG.md
@@ -1191,3 +1191,4 @@
 ### 1.0.3
 
 [Original release]
+