+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Master
 
+### 1.12.0
+
 - Upgrade palette@4.13.3; styled-components@4.x
 - Switch to use metaphysics v2 schema - justin/alloy
 - Fixes a bug where artist search results in consignments overlap with search bar - yuki24